Selectors, hover fix
