+++ apps/chat-e2e/src/ui/selectors/commonSelectors.ts
@@ -65,6 +65,17 @@ export const PopupSelectors = {
   popupContent: '[aria-label="popup-description"]',
 };
 
+export const UploadProgressDialogSelectors = {
+  container: '[role="dialog"][aria-modal="true"]:has-text("Uploading items")',
+  title: 'div:has-text("Uploading items")',
+  fileItem: '[data-qa="file-item"], .rounded.bg-layer-2',
+  fileName: '#name',
+  progressBar: '.bg-accent-primary',
+  progressBarContainer: '.rounded-full.bg-layer-1',
+  cancelButton: 'button[aria-label="Cancel"]',
+  closeButton: 'button[aria-label="Close dialog"]',
+};
+
 export const LoaderSelectors = {
   loaderContainer: '[role="status"]',
 };
