+++ apps/chat-e2e/src/ui/webElements/uploadProgressDialog.ts
@@ -1,4 +1,4 @@
-import { UploadProgressDialogSelectors } from '@/src/ui/selectors/uploadProgressDialogSelectors';
+import { UploadProgressDialogSelectors } from '@/src/ui/selectors';
 import { Popup } from '@/src/ui/webElements/common/popup';
 import { Locator, Page } from '@playwright/test';
